requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-openai
langchain-qdrant
langchain-community
qdrant-client
pypdf
python-docx

